@@ -8,8 +8,8 @@
     "start": "next start",
     "format:check": "prettier --check \"**/*.{md,html,mjml,json,yml,css,scss}\"",
     "format": "prettier --write \"**/*.{md,html,mjml,json,yml,css,scss,ts,tsx,js,jsx}\"",
-    "lint": "eslint 'app/javascript/**/*.{js,jsx,ts,tsx}' --max-warnings 0",
-    "lint:fix": "eslint 'app/javascript/**/*.{js,jsx,ts,tsx}' --fix",
+    "lint": "eslint '**/*.{ts,tsx}'",
+    "lint:fix": "eslint '**/*.{ts,tsx}' --fix",
     "link:bleu-ui": "npx yalc link @bleu-fi/ui"
   },
   "dependencies": {
@@ -68,8 +68,22 @@
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
+    "@typescript-eslint/eslint-plugin": "^6.21.0",
+    "@next/eslint-plugin-next": "^14.0.4",
+    "@typescript-eslint/parser": "^6.17.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-compat": "latest",
+    "eslint-plugin-graphql": "latest",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-prettier": "latest",
+    "eslint-plugin-react-hooks": "latest",
+    "eslint-plugin-simple-import-sort": "latest",
+    "eslint-plugin-tailwindcss": "^3.13.1",
+    "prettier": "^3.1.1",
+    "prettier-plugin-tailwindcss": "^0.5.10",
     "autoprefixer": "^10.0.1",
-    "eslint": "^8",
     "eslint-config-next": "14.1.0",
     "postcss": "^8",
     "tailwindcss": "^3.3.0",
