@@ -1,8 +1,10 @@
+import "./globals.css";
+
 import type { Metadata } from "next";
 import { Inter as FontSans } from "next/font/google";
-import "./globals.css";
-import { cn } from "#/lib/utils";
+
 import { RootLayout } from "#/components/RootLayout";
+import { cn } from "#/lib/utils";
 
 const fontSans = FontSans({
   subsets: ["latin"],
