@@ -1,4 +1,3 @@
-/* eslint-disable react/display-name */
 "use client";
 
 import { slateDarkA } from "@radix-ui/colors";
@@ -26,7 +25,7 @@ export const BaseInput = React.forwardRef<
     ref={ref}
     className={cn(
       "w-full selection:color-white box-border inline-flex h-[35px] appearance-none items-center justify-center rounded-[4px] bg-blue4 px-[10px] text-[15px] leading-none text-slate12 shadow-[0_0_0_1px] shadow-blue6 outline-none selection:bg-blue9 hover:shadow-[0_0_0_1px_black] focus:shadow-[0_0_0_2px_black] disabled:bg-blue1",
-      props.className,
+      props.className
     )}
   />
 ));
@@ -77,5 +76,5 @@ export const Input = React.forwardRef<HTMLInputElement, IInput>(
         )}
       </div>
     );
-  },
+  }
 );
