@@ -4,15 +4,16 @@ import SafeProvider from "@gnosis.pm/safe-apps-react-sdk";
 import Link from "next/link";
 import React from "react";
 
-import { ToastProvider } from "./Toast";
 import { NetworksContextProvider } from "@/app/contexts/networks";
+
+import { ToastProvider } from "./Toast";
 
 export function RootLayout({ children }: React.PropsWithChildren) {
   return (
     <SafeProvider loader={<SafeLoader />}>
       <ToastProvider>
         <NetworksContextProvider>
-          <div className="h-full w-full bg-blue1">{children}</div>
+          <div className="size-full bg-blue1">{children}</div>
         </NetworksContextProvider>
       </ToastProvider>
     </SafeProvider>
@@ -21,7 +22,7 @@ export function RootLayout({ children }: React.PropsWithChildren) {
 
 function SafeLoader() {
   return (
-    <div className="flex h-full w-full flex-col  justify-center items-center rounded-3xl px-12 py-16 md:py-20 text-slate12">
+    <div className="flex size-full flex-col justify-center  items-center rounded-3xl px-12 py-16 md:py-20 text-slate12">
       <div className="text-center text-3xl">This is a Safe (wallet) App</div>
       <p className="text-xl">
         To access please use this
