@@ -1,6 +1,7 @@
+import Image from "next/image";
+
 import { cowTokenList } from "#/lib/cowTokenList";
 import { formatNumber, truncateAddress } from "#/lib/utils";
-import Image from "next/image";
 
 export function TokenInfo({
   symbol,
